Added joystick mode, debug command now toggle
When the joystick mode flag is enabled (general command 23), the
controller will ignore all commands except those that are relevant to
operating the joystick in the mobile app. This is to avoid unexpected
behavior that might occur from corrupted Bluetooth packets being
received as command other than what was intended.

The debug command had required explicit setting of the LED on or off.
This has been changed to a simple toggle with a data length of 0.