Joystick chng, auto set spd, fixed max step rate query
When in joystick mode, no simple move commands are required to move the
motor. When the continuous motor speed is set to any non-zero value, the
controller will automatically start a move. When turning on joystick
mode, it will change the continuous speed to 0 to prevent runaway
motors.

When "send to home", "send to start", etc commands are called, the
controller will automatically set the current speed to the default
maximum speed. This prevents a condition where the motors will not move
because they had been previously set to 0 speed.

The max motor step rate query was previously always responding with null
values. This has been fixed.

The "set start here" and "set stop here" commands have been changed to
motor commands 228 and 229, respectively. They are for debugging
purposes only and will be removed later.