Major update
--Added watchdog mode and watchdog mode queries to acceptable joystick
mode commands
--Enhanced pause mode handling
--Watchdog mode automatically disabled when exiting joystick mode
--Added general command 25 to send all motors to program start position
simultaneously
--Added general commands 26 and 27 to set all motors program start and
stop points simultaneously, respectively
--Modified run time status query to return the elapsed running time
after a program has stopped rather than re-setting it to 0
--Added general query command 125 to return the total program time based
upon currently set program parameters
--Added general query command 126 to return whether the program is
complete. It will return true the first time it is called after a
program completes and false thereafter.
--Modified motor send-to-start handling
--Pressing the e-stop button during a program stops the program instead
of pausing it. The app is not aware of this change of state, so stopping
altogether is safer.

--Added lots of debug serial USBSerial prints