Fixed float/fixed point responses
The float response handler on in the firmware doesn't seem to work, so
responses from the controller that would normally transmit floats should
be multiplied by 100 and transmitted as an unsigned long. When they are
received on the master side, they can be divided by 100 to give a fixed
point number with two decimal places. Before this fix, the floats
weren't being multiplied or being cast as unsigned longs.