Removed direction setting for joystick speed 0
If the speed was set to 0, the motor direction was previously set to 1.
This caused the controller to momentarily, but erroneously think there
was backlash that needed to be taken up in cases where the motor was
sent in reverse twice in a row. This caused the motor position indicator
on Graffik to hop around.