Removed backlash take-up from key frame program start
By having the backlash take-up at the beginning of the program, it took
a long time before the controller was free to accept serial commands
again. If any commands were sent during this window, it could cause the
controller to lock and/or the communications link to fail.

The backlash take-up has been extracted as a separate serial command so
developers can be aware that it is occurring and build in a special wait
or exception to sub 300ms coms timeouts.