Fixed delay flag and time reporting for KF mode
- The delay flag was not being set in KF mode, so the run status report
was incorrect during the delay segment of the program.
- The total time reported for KF mode was based upon a static var that
was set once at the beginning of the program and was not recalculated,
only reported during the program. For some reason, this value was
getting changed to an arbitrarily large value. It is now recalculated on
every function call, since it only increases the reporting time from 4us
to 8us.