-Redundancy check to make sure motors are at start before starting move
Re: @milapse 8:49 AM Sounds good, one thing to double check on the
PS3.. I was testing this weekend and tried hitting the emergency stop
while it was returning to a home position. When I hit 'start' again it
seemed to be starting the program from the spot where I hit the e-stop.
So perhaps you'll need to do a quick check when the user taps start
that the current motor positions are equal to the start point.