Motor detection works with higher supply voltage
The current threshold for motor detection was based upon an assumed 12V
supply. The actual current drawn by a single motor is lower given a
higher supply, though, so this can cause the function to not detect
motors that are in fact connected. This has been accounted for now by
setting a power threshold and calculating the current threshold based
upon the voltage pin reading.