Updated to version 77
There was a strange instability problem with v0.76 that may have been
related to a bug present at the time of the hex file compilation or a
corrupted hex file. I reset the OMLibrary files to the latest version
available on the repo (though the changes present locally didn't seem
like they had existed at the time of the last hex file compilation) and
re-compiled the hex file, which seems to have solved the problem. I've
incremented the version number so as to avoid ambiguity on the web
updater.